move Pay editor form to @planx/components/Pay/Editor.tsx
